generateChallenge to auth utils
